Game Engine and Main class
Main class constructs and starts a Game object.

Game runs the game using a simple engine. Start method starts the
threads, update method update the objects of the game, and renderize
method render then on Screen. All of them are called recursively inside
the run method, with a 60frames per second rate,it checks if the Game is
still runing before update (isRuning property).

Makefile was designed to compile and run the classes.