Screen class to start JFrame, canvas and screen
Screen class defines a screen object with a JFrame and a canvas. It
draws inside a layer, wich is a bufferedImage object. It creates a
bufferStrategy in canvas to load the image on screen using buffers, in
order to optimize the renderize process.

Now, background it's not loading, only JFrame and canvas, because you've
got to call drawBackground Screen method on the renderize Game method to
render.