render player image on screen
Spritesheet class tries to get a spritesheet given a string path.
It uses getSprite() to cut a piece of a spritesheet and return it to a
bufferedImage that will load it on screen.

drawPlayer() Screen method loads the bufferedImage image on a Graphics2D
object and with bufferStrategy bs.show() it'll be drawn on the screen.