Player movement and animation
Implements Player movements and animations. Keylistener and KeyEvent are
used to listen if the directional keys are pressed and to modify
player's position (update method) and animate (renderize and player
animate methods).

Method animatePlayer sets the players image during the animation based
on four booleans to determine wich direction the player is moving and
animate correctly.