Loading map methods (scratch)
Logic to load map.png into game screen. Use map.png as a template,
assigning each pixel color in the image to a sprite in spritesheet.png.

The method loadMap() loads an array of pixel colors of map.png, if the
color assigned is red (0xFFFF0000) it'll print a message on screen.
Someone could use that logic to load spritesheets on screen based on
color pixels on map.png, it'll identify the possition of static objects
(tiles) such as floor, grass, etc.