Player colision with the wall
Method isFree() from Screen class verifies if the position that Players
wants to move 'is free', it means 'check if there is a wall'. It returns
true if there's no obstacle in Player's path, and false if there is.

The Method is Free uses the center of the Player's spritesheet, the
x and y coordinates of the center, in the game map on the screen. It
maps those coordinates into a pixel map used by Map class to build the
game screen.

The color of every pixel in the pixel map means one tile into game
screen, Map class converts that mapping every pixel into a tile on the
screen. So, that map can be used to identify an object on screen based
on it's color pixel. For a wall it is WHITE (0xFFFFFFFF).